update: split CustomHeader into two component
